+++ app/src/main/java/fr/uge/plutus/data/model/Wallet.kt
@@ -6,6 +6,6 @@ import androidx.room.PrimaryKey
 @Entity
 data class Wallet(
     @PrimaryKey(autoGenerate = true)
-    val walletId: Int?,
+    val walletId: Long?,
     val name: String
 )
