+++ app/src/main/java/fr/uge/plutus/storage/StorageState.kt
@@ -2,5 +2,5 @@ package fr.uge.plutus.storage
 
 data class StorageState(
     val dark: Boolean? = null,
-    val wallet: Int? = null
+    val wallet: Long? = null
 )
